# tui/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "skene",
    "description": "Backend server for the skene analysis engine.",
    "version": "0.5.1"
  },
  "paths": {
...
    "/journey/analyse": {
      "post": {
        "tags": [
          "journey"
        ],
        "summary": "Analyse",
        "description": "Start the canned journey analysis in a fresh session.\n\nThe run proceeds async: watch ``/event`` for progress; the result lands\nas an ``artifact`` part (and as ``journey.yaml`` in the workspace).",
        "operationId": "analyse_journey_analyse_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "x-skene-directory",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Skene-Directory",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyAnalyseRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyAnalyseAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/journey": {
      "get": {
        "tags": [
          "journey"
        ],
        "summary": "Get Journey",
        "description": "Latest ``journey.yaml`` for the workspace, parsed to JSON.",
        "operationId": "get_journey_journey_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "x-skene-directory",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Skene-Directory",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Journey Journey Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JourneyAnalyseAccepted": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "JourneyAnalyseAccepted",
        "description": "Ack for ``POST /journey/analyse``: the run proceeds async in this session."
      },
      "JourneyAnalyseRequest": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string",
            "nullable": true
          },
          "schemaDir": {
            "title": "Schemadir",
            "type": "string",
            "nullable": true
          },
          "dbUrl": {
            "title": "Dburl",
            "type": "string",
            "nullable": true
          },
          "productName": {
            "title": "Productname",
            "type": "string",
            "nullable": true
          },
          "output": {
            "title": "Output",
            "type": "string",
            "nullable": true
          },
          "schemaMaxTurns": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Schemamaxturns",
            "default": 150
          },
          "codeMaxTurns": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Codemaxturns",
            "default": 200
          },
          "classifyConcurrency": {
            "type": "integer",
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Classifyconcurrency",
            "default": 8
          },
          "specialize": {
            "type": "boolean",
            "title": "Specialize",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "JourneyAnalyseRequest",
        "description": "Body of ``POST /journey/analyse`` \u2014 the v1-compat canned analysis.\n\nMirrors the ``skene analyse-journey`` CLI flags. ``db_url`` is used for\nlive introspection only and is never persisted or echoed back \u2014 the\nstored session records a redacted display form."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}